.windsurf/skills/plan/.shared/scripts/include/pyproject.toml
[project]
name = "plan-cli"
version = "0.1.0"
description = "CLI for plan artifact management"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
